// src/frontend/webapp/modilist/src/services/swagger/api/models/address-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ModilistAPI
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
/**
 * 
 * @export
 * @interface AddressDTO
 */
export interface AddressDTO {
    /**
     * 
     * @type {number}
     * @memberof AddressDTO
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof AddressDTO
     */
    accountId?: string;
    /**
     * 
     * @type {string}
     * @memberof AddressDTO
     */
    name?: string;
    /**
     * 
     * @type {string}
     * @memberof AddressDTO
     */
    firstName?: string;
    /**
     * 
     * @type {string}
     * @memberof AddressDTO
     */
    lastName?: string;
    /**
     * 
     * @type {string}
     * @memberof AddressDTO
     */
    phone?: string;
    /**
     * 
     * @type {string}
     * @memberof AddressDTO
     */
    zipCode?: string;
    /**
     * 
     * @type {string}
     * @memberof AddressDTO
     */
    city?: string;
    /**
     * 
     * @type {string}
     * @memberof AddressDTO
     */
    district?: string;
    /**
     * 
     * @type {string}
     * @memberof AddressDTO
     */
    fullAddress?: string;
    /**
     * 
     * @type {boolean}
     * @memberof AddressDTO
     */
    isDefault?: boolean;
}